requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
requests==2.31.0
yt-dlp>=2024.12.0
mutagen>=1.47.0
psycopg2-binary>=2.9.9
